Extract IChecksum to IFileCache
