+++ src/SonarScanner.MSBuild.PreProcessor/PreprocessorObjectFactory.cs
@@ -103,8 +103,8 @@ public class PreprocessorObjectFactory : IPreprocessorObjectFactory
     public IJreResolver CreateJreResolver(ISonarWebServer server, string sonarUserHome)
     {
         var filePermissionsWrapper = new FilePermissionsWrapper(new OperatingSystemProvider(FileWrapper.Instance, logger));
-        var fileCache = new FileCache(DirectoryWrapper.Instance, FileWrapper.Instance, sonarUserHome);
-        var cache = new JreCache(logger, fileCache, DirectoryWrapper.Instance, FileWrapper.Instance, ChecksumSha256.Instance, UnpackerFactory.Instance, filePermissionsWrapper);
+        var fileCache = new FileCache(logger, DirectoryWrapper.Instance, FileWrapper.Instance, ChecksumSha256.Instance, sonarUserHome);
+        var cache = new JreCache(logger, fileCache, DirectoryWrapper.Instance, FileWrapper.Instance, UnpackerFactory.Instance, filePermissionsWrapper);
         return new JreResolver(server, cache, logger);
     }
